Add files via upload
This project is a simple BMI calculator that uses Bootstrap to make the UI visually pleasing. The web app takes user input for gender, weight, and height. Then, the user's BMI is calculated and returned to the user along with a category, such as underweight. Unfortunately, this file is too large to have a preview. If you would like to view my code, please download and extract the files and use Visual Studio 2022 to open the files. Enjoy!