Update BMICalculatorWithBootstrapUI.ZIP
This program is a BMI calculator that utilizes a Bootstrap UI to enhance the overall appearance. The program takes input from the user: gender, height, and weight. Then, the program calculates the user's BMI and returns it to the user along with a fitness category (e.g., underweight).  Unfortunately, this file is too large to generate a preview. If you would like to view my code, please download the ZIP file and open it in Visual Studio 2022. Enjoy!